moved pokemon list related functions to new files
add file displayList.js
add file pokemonList.js
